refactor: move ui theme provider and chat reasoning
