+++ ui/src/components/right-panel/agent/chat-reasoning.tsx
